backend/requirements.txt
Django~=2.2.4
Pillow
djangorestframework~=3.11.0